hot fix

+++ Fish Tycoon/Assets/Scripts/BoatBehavior.cs
@@ -60,9 +60,9 @@ public class BoatBehavior : MonoBehaviour {
 
 	void Move() {
         if (!docked) {
-            int overcapacity = 1.0f;
+            float overcapacity = 1.0f;
 
-            overcapacity += 0.1 * Mathf.Max (0, (capacity - maxCapaciity));
+            overcapacity += 0.1f * Mathf.Max (0, (capacity - maxCapaciity));
             fuel = fuel - fuelConsumptionRate * Time.deltaTime;
         }
